+++ utilities/persistent_cache/block_cache_tier_file.h
@@ -187,7 +187,7 @@ class WriteableCacheFile : public RandomAccessCacheFile {
 
   // create file on disk
   bool Create(const bool enable_direct_writes, const bool enable_direct_reads,
-              const size_t page_alignment);
+              const uint32_t page_alignment);
 
   // read data from logical file
   bool Read(const LBA& lba, Slice* key, Slice* block, char* scratch) override {
@@ -213,7 +213,7 @@ class WriteableCacheFile : public RandomAccessCacheFile {
 
   bool ReadBuffer(const LBA& lba, Slice* key, Slice* block, char* scratch);
   bool ReadBuffer(const LBA& lba, char* data);
-  bool PadZeros(const size_t size);
+  bool PadZeros(const uint32_t size);
   bool ExpandBuffer(const size_t size);
   void DispatchBuffer();
   void BufferWriteDone();
@@ -247,7 +247,7 @@ class WriteableCacheFile : public RandomAccessCacheFile {
   size_t pending_ios_ = 0;               // Number of ios to disk in-progress
   bool enable_direct_reads_ = false;     // Should we enable direct reads
                                          // when reading from disk
-  size_t page_alignment_ = 0;            // Align writes with this page size
+  uint32_t page_alignment_ = 0;          // Align writes with this page size
                                          // (disabled if 0)
 };
